refactor: events -> messaging 패키지명 통일, producer/consumer 패키지 분리
